Autovote & Respawn Hotfix? (#28)
* AUTOMATE ALL!

* Respawn Hotfix

+++ Content.Shared/CCVar/CCVars.GhostRespawn.cs
@@ -4,8 +4,8 @@ namespace Content.Shared.CCVar;
 
 public sealed partial class CCVars
 {
-    public static readonly CVarDef<int> GhostRespawnTime =
-        CVarDef.Create("ghost.respawn_time", 15, CVar.SERVERONLY);
+    public static readonly CVarDef<double> GhostRespawnTime =
+        CVarDef.Create("ghost.respawn_time", 15d, CVar.SERVERONLY);
 
     public static readonly CVarDef<int> GhostRespawnMaxPlayers =
         CVarDef.Create("ghost.respawn_max_players", 40, CVar.SERVERONLY);
